requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
aiofiles
google-generativeai
